+++ CHANGELOG.md
@@ -59,6 +59,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - conditions `biome`, `effect`, `ride`, `random` now support variables
 - `sync` variable offering identical evaluation as `eval` but forcing the execution on the main thread 
 - native `menu` io implementation which sets the player's speed to zero instead mounting it on an arrow (MC 1.21.4+)
+- support for `itemModel` for the totem NotifyIO (MC 1.21.4+)
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations